refactor: organize components by feature
