+++ src/app/app/[org_id]/[team_id]/posts/page.tsx
@@ -4,9 +4,9 @@ import { HydrationBoundary, dehydrate } from '@tanstack/react-query';
 import { getQueryClient } from '@/components/providers/get-query-client';
 import { api } from '@/queries';
 import { postsKeys } from '@/queries/keys';
-import PostsTable from '@/components/posts-table';
+import PostsTable from '@/features/posts/posts-table';
 import { DashboardHeader } from '@/components/dashboard-header';
-import { CreatePostModal } from '@/components/create-post-modal';
+import { CreatePostModal } from '@/features/posts/create-post-modal';
 import { DashboardShell } from '@/components/dashboard-shell';
 
 export default async function Page({
