+++ src/app/app/[org_id]/members/page.tsx
@@ -1,7 +1,7 @@
-import MembersTable from '@/components/members-table'
+import MembersTable from '@/features/members/members-table'
 import { createClient } from '@/lib/supabase/server'
 import { redirect } from 'next/navigation'
-import { InviteMemberModal } from '@/components/invite-member-modal'
+import { InviteMemberModal } from '@/features/invitations/invite-member-modal'
 import { DashboardHeader } from '@/components/dashboard-header'
 import { DashboardShell } from '@/components/dashboard-shell'
 import { HydrationBoundary, dehydrate } from '@tanstack/react-query'
