+++ src/app/app/[org_id]/settings/page.tsx
@@ -1,7 +1,7 @@
 import { redirect } from 'next/navigation'
 import { createClient } from '@/lib/supabase/server'
-import UpdateOrganizationForm from '@/components/update-organization-form'
-import { DeleteOrganizationButton } from '@/components/delete-organization-button'
+import UpdateOrganizationForm from '@/features/organizations/update-organization-form'
+import { DeleteOrganizationButton } from '@/features/organizations/delete-organization-button'
 import { HydrationBoundary, dehydrate } from '@tanstack/react-query'
 import { getQueryClient } from '@/components/providers/get-query-client'
 import { api } from '@/queries'
