+++ src/app/auth/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from '@/components/login-form'
+import { LoginForm } from '@/features/auth/login-form'
 
 export default function Page() {
   return (
